Module Ipaddr_unix.V6

val to_inet_addr : Ipaddr.V6.t -> Unix.inet_addr

to_inet_addr ip is the Unix.inet_addr equivalent of the IPv6 address ip.

val of_inet_addr_exn : Unix.inet_addr -> Ipaddr.V6.t

of_inet_addr_exn ip is the Ipaddr.t equivalent of the Unix.inet_addr ip IPv6 address. Raises Ipaddr.Parse_error if ip is not a valid representation of an IPv6 address.

val of_inet_addr : Unix.inet_addr -> Ipaddr.V6.t option

Same as of_inet_addr_exn but returns an option type instead of raising an exception.